Fix tests (port hyperlane unit test)
