+++ chainweb.cabal
@@ -241,7 +241,9 @@ library
         , Chainweb.VerifierPlugin
         , Chainweb.VerifierPlugin.Allow
         , Chainweb.VerifierPlugin.Hyperlane.Announcement
+        , Chainweb.VerifierPlugin.Hyperlane.Binary
         , Chainweb.VerifierPlugin.Hyperlane.Message
+        , Chainweb.VerifierPlugin.Hyperlane.Utils
         , Chainweb.Version
         , Chainweb.Version.Development
         , Chainweb.Version.Guards
@@ -303,7 +305,6 @@ library
         , Chainweb.Pact.RestAPI.SPV
         , Chainweb.Pact.RestAPI.Server
         , Chainweb.Pact.SPV
-        , Chainweb.Pact.SPV.Hyperlane.Binary
         , Chainweb.Pact.Service.BlockValidation
         , Chainweb.Pact.Service.PactInProcApi
         , Chainweb.Pact.Service.PactQueue
@@ -453,7 +454,7 @@ library
 -- -------------------------------------------------------------------------- --
 
 test-suite chainweb-tests
-    import: debugging-flags
+    import: warning-flags, debugging-flags
     default-language: Haskell2010
     ghc-options:
         -threaded
@@ -485,6 +486,8 @@ test-suite chainweb-tests
         Chainweb.Test.Pact.GrandHash
         Chainweb.Test.Pact.PactMultiChainTest
         Chainweb.Test.Pact.VerifierPluginTest
+        Chainweb.Test.Pact.VerifierPluginTest.Transaction
+        Chainweb.Test.Pact.VerifierPluginTest.Unit
         Chainweb.Test.Pact.PactSingleChainTest
         Chainweb.Test.P2P.Peer.BootstrapConfig
         Chainweb.Test.Pact.DbCacheTest
