+++ test/ChainwebTests.hs
@@ -89,10 +89,10 @@ main = do
             liftIO $ defaultMainWithIngredients (consoleAndJsonReporter : defaultIngredients)
                 $ adjustOption adj
                 $ testGroup "Chainweb Tests"
-                $ [pactTestSuite rdb]
-                -- : mempoolTestSuite db h0
-                -- : rosettaTestSuite rdb
-                -- : suite rdb
+                $ pactTestSuite rdb
+                : mempoolTestSuite db h0
+                : rosettaTestSuite rdb
+                : suite rdb
   where
     adj NoTimeout = Timeout (1_000_000 * 60 * 10) "10m"
     adj x = x
@@ -103,19 +103,19 @@ mempoolTestSuite db genesisBlock = testGroup "Mempool Consensus Tests"
 
 pactTestSuite :: RocksDb -> TestTree
 pactTestSuite rdb = testGroup "Chainweb-Pact Tests"
-    [ -- Chainweb.Test.Pact.PactExec.tests
-    -- , Chainweb.Test.Pact.DbCacheTest.tests
-    -- , Chainweb.Test.Pact.Checkpointer.tests
-        -- Chainweb.Test.Pact.PactMultiChainTest.tests
-        Chainweb.Test.Pact.VerifierPluginTest.tests
-    -- , Chainweb.Test.Pact.PactSingleChainTest.tests rdb
-    -- , Chainweb.Test.Pact.RemotePactTest.tests rdb
-    -- , Chainweb.Test.Pact.PactReplay.tests rdb
-    -- , Chainweb.Test.Pact.ModuleCacheOnRestart.tests rdb
-    -- , Chainweb.Test.Pact.TTL.tests rdb
-    -- , Chainweb.Test.Pact.RewardsTest.tests
-    -- , Chainweb.Test.Pact.NoCoinbase.tests
-    -- , Chainweb.Test.Pact.GrandHash.tests
+    [ Chainweb.Test.Pact.PactExec.tests
+    , Chainweb.Test.Pact.DbCacheTest.tests
+    , Chainweb.Test.Pact.Checkpointer.tests
+    , Chainweb.Test.Pact.PactMultiChainTest.tests
+    , Chainweb.Test.Pact.VerifierPluginTest.tests
+    , Chainweb.Test.Pact.PactSingleChainTest.tests rdb
+    , Chainweb.Test.Pact.RemotePactTest.tests rdb
+    , Chainweb.Test.Pact.PactReplay.tests rdb
+    , Chainweb.Test.Pact.ModuleCacheOnRestart.tests rdb
+    , Chainweb.Test.Pact.TTL.tests rdb
+    , Chainweb.Test.Pact.RewardsTest.tests
+    , Chainweb.Test.Pact.NoCoinbase.tests
+    , Chainweb.Test.Pact.GrandHash.tests
     ]
 
 rosettaTestSuite :: RocksDb -> TestTree
